move RejectApplicationForm in it's own file
