+++ front/src/app/components/admin/establishments/DiscussionManageContent.tsx
@@ -6,10 +6,7 @@ import ButtonsGroup from "@codegouvfr/react-dsfr/ButtonsGroup";
 import { Card } from "@codegouvfr/react-dsfr/Card";
 import Input from "@codegouvfr/react-dsfr/Input";
 import { createModal } from "@codegouvfr/react-dsfr/Modal";
-import Select, { type SelectProps } from "@codegouvfr/react-dsfr/SelectNext";
-import { zodResolver } from "@hookform/resolvers/zod";
-import { renderContent } from "html-templates/src/components/email";
-import { useEffect, useState } from "react";
+import { useEffect } from "react";
 import {
   ButtonWithSubMenu,
   DiscussionMeta,
@@ -17,32 +14,26 @@ import {
   Loader,
 } from "react-design-system";
 import { createPortal } from "react-dom";
-import { useForm } from "react-hook-form";
 import { useDispatch } from "react-redux";
 import {
   type DiscussionDisplayStatus,
   type DiscussionId,
   type DiscussionReadDto,
   type Email,
-  type ExchangeFromDashboard,
-  type RejectDiscussionAndSendNotificationParam,
-  type RejectionKind,
   type WithDiscussionId,
-  type WithDiscussionRejection,
+  type ExchangeFromDashboard,
   addressDtoToString,
   createOpaqueEmail,
-  discussionRejectionSchema,
   domElementIds,
   escapeHtml,
   exchangeMessageFromDashboardSchema,
   getDiscussionDisplayStatus,
-  rejectDiscussionEmailParams,
   toDisplayedDate,
 } from "shared";
+import { RejectApplicationForm } from "src/app/components/admin/establishments/RejectApplicationForm";
 import type { ConventionPresentation } from "src/app/components/forms/convention/conventionHelpers";
 import { useDiscussion } from "src/app/hooks/discussion.hooks";
 import { useFeedbackEventCallback } from "src/app/hooks/feedback.hooks";
-import { makeFieldError } from "src/app/hooks/formContents.hooks";
 import { useAppSelector } from "src/app/hooks/reduxHooks";
 import {
   getConventionInitialValuesFromUrl,
@@ -438,212 +429,3 @@ const makeConventionFromDiscussion = ({
   immersionAppellation: discussion.appellation,
   immersionAddress: addressDtoToString(discussion.address),
 });
-
-const RejectApplicationForm = ({
-  discussion,
-}: {
-  discussion: DiscussionReadDto;
-}): JSX.Element => {
-  const { register, watch, handleSubmit, formState } =
-    useForm<WithDiscussionRejection>({
-      resolver: zodResolver(discussionRejectionSchema),
-    });
-  const inclusionConnectedJwt = useAppSelector(
-    authSelectors.inclusionConnectToken,
-  );
-  const getFieldError = makeFieldError(formState);
-  const dispatch = useDispatch();
-  const watchedFormValues = watch();
-  const [rejectParams, setRejectParams] =
-    useState<RejectDiscussionAndSendNotificationParam | null>(null);
-  const rejectionKindOptions: SelectProps.Option<RejectionKind>[] = [
-    {
-      label:
-        "Je ne suis pas en capacité d'aider le candidat dans son projet professionnel",
-      value: "UNABLE_TO_HELP",
-    },
-    {
-      label:
-        "Je traverse une période chargée et je n'ai pas le temps d'accueillir une immersion",
-      value: "NO_TIME",
-    },
-    {
-      label: "Autre",
-      value: "OTHER",
-    },
-  ];
-  if (!inclusionConnectedJwt) throw new Error("No jwt found");
-  if (rejectParams) {
-    const { htmlContent, subject } = rejectDiscussionEmailParams(
-      rejectParams,
-      discussion,
-    );
-    return (
-      <>
-        <p>
-          Voici l'email qui sera envoyé au candidat si vous rejetez sa
-          candidature :
-        </p>
-        <Card
-          title={`Sujet : ${subject}`}
-          className={fr.cx("fr-mb-2w")}
-          desc={
-            <div
-              dangerouslySetInnerHTML={{
-                __html:
-                  renderContent(htmlContent, { wrapInTable: false }) ?? "",
-              }}
-            />
-          }
-        />
-        <ButtonsGroup
-          buttons={[
-            {
-              id: domElementIds.establishmentDashboard.discussion
-                .rejectApplicationCancelButton,
-              priority: "secondary",
-              children: "Annuler",
-              type: "button",
-              onClick: () => setRejectParams(null),
-            },
-            {
-              id: domElementIds.establishmentDashboard.discussion
-                .rejectApplicationSubmitButton,
-              priority: "primary",
-              children: "Rejeter la candidature et notifier le candidat",
-              onClick: () =>
-                dispatch(
-                  discussionSlice.actions.updateDiscussionStatusRequested({
-                    status: "REJECTED",
-                    feedbackTopic: "dashboard-discussion-rejection",
-                    ...rejectParams,
-                    jwt: inclusionConnectedJwt,
-                  }),
-                ),
-            },
-          ]}
-          inlineLayoutWhen="always"
-        />
-      </>
-    );
-  }
-
-  return (
-    <form
-      onSubmit={handleSubmit((values) =>
-        setRejectParams({ ...values, discussionId: discussion.id }),
-      )}
-    >
-      <Select
-        label="Pour quel motif souhaitez-vous refuser cette candidature ?"
-        nativeSelectProps={{
-          id: domElementIds.establishmentDashboard.discussion
-            .rejectApplicationJustificationKindInput,
-          ...register("rejectionKind"),
-        }}
-        options={rejectionKindOptions}
-        {...getFieldError("rejectionKind")}
-      />
-      {watchedFormValues.rejectionKind === "OTHER" && (
-        <Input
-          textArea
-          label="Précisez"
-          nativeTextAreaProps={{
-            id: domElementIds.establishmentDashboard.discussion
-              .rejectApplicationJustificationReasonInput,
-            ...register("rejectionReason"),
-          }}
-          {...getFieldError("rejectionReason")}
-        />
-      )}
-      <ButtonsGroup
-        buttons={[
-          {
-            id: domElementIds.establishmentDashboard.discussion
-              .rejectApplicationCancelButton,
-            priority: "secondary",
-            children: "Annuler",
-            type: "button",
-          },
-          {
-            id: domElementIds.establishmentDashboard.discussion
-              .rejectApplicationSubmitPreviewButton,
-            priority: "primary",
-            children: "Prévisualiser et envoyer",
-          },
-        ]}
-        inlineLayoutWhen="always"
-      />
-    </form>
-  );
-};
-
-const DiscussionEchangeMessageForm = ({
-  discussionId,
-}: {
-  discussionId: DiscussionId;
-}) => {
-  const { register, handleSubmit, formState } = useForm<ExchangeFromDashboard>({
-    resolver: zodResolver(exchangeMessageFromDashboardSchema),
-    defaultValues: {
-      message: "",
-    },
-  });
-  const getFieldError = makeFieldError(formState);
-  const dispatch = useDispatch();
-  const inclusionConnectedJwt = useAppSelector(
-    authSelectors.inclusionConnectToken,
-  );
-
-  const onSubmit = (data: ExchangeFromDashboard) => {
-    if (inclusionConnectedJwt) {
-      dispatch(
-        discussionSlice.actions.sendExchangeRequested({
-          exchangeData: {
-            jwt: inclusionConnectedJwt,
-            discussionId,
-            message: data.message,
-          },
-          feedbackTopic: "establishment-dashboard-discussion-send-message",
-        }),
-      );
-    }
-  };
-
-  return (
-    <form onSubmit={handleSubmit(onSubmit)} className={fr.cx("fr-mb-4w")}>
-      <Feedback
-        topics={["establishment-dashboard-discussion-send-message"]}
-        className={fr.cx("fr-mb-2w")}
-        closable
-      />
-      <input type="hidden" {...register("discussionId")} value={discussionId} />
-      <Input
-        textArea
-        label="Répondre au candidat"
-        nativeTextAreaProps={{
-          id: domElementIds.establishmentDashboard.discussion.sendMessageInput,
-          rows: 5,
-          placeholder: "Rédigez votre message ici...",
-          ...register("message", {
-            setValueAs: escapeHtml,
-          }),
-        }}
-        {...getFieldError("message")}
-      />
-      <div className={fr.cx("fr-mt-2w", "fr-mb-4w")}>
-        <Button
-          id={
-            domElementIds.establishmentDashboard.discussion
-              .sendMessageSubmitButton
-          }
-          type="submit"
-          disabled={formState.isSubmitting}
-          size="small"
-        >
-          Envoyer un message
-        </Button>
-      </div>
-    </form>
-  );
-};
